Fix docker composer
I have noticed that PYTHONPATH needs to be updated with the current path,
in order to properly start web_server, although reddit_daemon starts ok
even if PYTHONPATH is not updated.

I have also hardcoded in web_server the IP address of the mongodb container,
because the IP address of the mongodb container does not appear in hosts file of
the web_server container, even if the web_server container is linked to the mongodb
container, in docker-compose.yml.

I did the same hardcoding workaround in the reddit_daemon too, for the same reason,
in a previous commit.

Signed-off-by: andanicolae <nicolae.anda@gmail.com>